# Improve performance by using `Vec<T>` instead of `Vec<Vec<T>>` for 2D array