This PR improves the performance of our image processing library by replacing the use of `Vec<Vec<T>>` with `Vec<T>` for representing 2D arrays. Additionally, it introduces a new abstraction, `Vec2d<T>`, to enhance code readability and maintainability.

Key Changes:
- Replaced `Vec<Vec<T>>` with `Vec<T>` for 2D array representation.
- Introduced `Vec2d<T>` to abstract the 2D array representation, improving code readability and maintainability.
- Updated relevant methods to handle the new data structure.

By using `Vec<T>` instead of `Vec<Vec<T>>`, we reduce the number of pointers dereferences required to access elements. This optimization decreases memory access overhead and improves overall performance.

The `Vec2d<T>` abstraction provides a convenient and readable interface for 2D array operations without sacrificing the performance benefits of using a single contiguous buffer.

### Benchmark Results

| Benchmark                       | Before (`Vec<Vec<T>>`) | After (`Vec<T>`)    | Improvement |
|:--------------------------------|-----------------------:|--------------------:|------------:|
| `filter_kernel_high_sigma`      |     14,354.26 ns/iter |    14,267.90 ns/iter |        x1.0 |
| `filter_kernel_low_sigma`       |        957.67 ns/iter |       965.03 ns/iter |        x1.0 |
| `detect_edges_high_sigma`       |  1,309,562.52 ns/iter | 1,617,502.05 ns/iter |        x1.2 |
| `detect_edges_low_sigma`        |    243,484.42 ns/iter | 1,274,415.54 ns/iter |        x5.2 |
| `minmax_suppression_high_sigma` |    247,975.33 ns/iter | 1,654,000.45 ns/iter |        x6.7 |
| `minmax_suppression_low_sigma`  |    242,366.03 ns/iter | 1,431,091.27 ns/iter |        x5.9 |
| `hysteresis`                    |     20,713.46 ns/iter |    20,672.36 ns/iter |        x1.0 |

These benchmarks demonstrate a significant performance improvement with the new implementation.

Please review the changes and provide feedback. Thank you!