@gifnksm 2024-07-29
It has been observed that this change results in performance degradation for large images (e.g., 674x274). I don't know the cause of this issue currently. I would need additional investigation to identify and address the underlying problem.