@gifnksm 2024-07-29
I completely misunderstood the benchmark results...